fix: chatbot
