+++ src/utilities/typeGuards/index.ts
@@ -1,0 +1,5 @@
+export {
+	isNavigateToProjectMessage,
+	isNavigateToConnectionMessage,
+	isVarUpdatedMessage,
+} from "./iframeCommunication.typeGuards";
